dataoppsamling/requirements.txt
appdirs==1.4.4
azure-common==1.1.25
azure-core==1.8.1
azure-identity==1.4.0
azure-keyvault-secrets==4.2.0
black==20.8b1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
Flask==1.1.2
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msal==1.5.0
msal-extensions==0.2.2
msrest==0.6.19
mypy-extensions==0.4.3
mysql-connector-python==8.0.21
mysqlclient==2.0.1
oauthlib==3.1.0
pathspec==0.8.0
portalocker==1.7.1
protobuf==3.13.0
pycparser==2.20
PyJWT==1.7.1
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
